Requests - upload path, link, specs
- Uses strong parameters in requestcontroller
- Gets user from current_user
- Consistently uses "external_id" instead of "mirlyn_id"

- Use 'create' instead of 'build' and pass parameters to create instead
of new. This more closely matches the general expectation for a factory
that saves the thing it makes.

- Continues to just pass the request parameters through the factory to
the object initializer. I tried some things with keyword args, but the
way Rails handles parameters vs. the Ruby keyword args makes it messy to
marshal/unmarshal things that just end up getting passed through to the
object constructor anyway.