Error handling for BagMoveJobs
Rescues StandardError, logs the error in the queueitem, and re-throws
the error.  If the queue item can't be saved, the original exception
could be lost, but at least there would still be an error logged for the
job in the Rails error log (just not the original error).

The example upload CLI also now prints the returned queue item (in case
of error) or the finished bag (in case of success). It also prints the
detail for internal server errors rather than just throwing an exception.