Collapse Request into Bag
This was a huge change, despite their functionality being so similar.
We did this due to the huge complexity penalty we were paying for
very little gain.

This commit is mostly to ensure that the tests are still there and
that they pass.  There are likely still edges that need fixing,
especially in the /lib classes.