PFDR-67: validates chipmunk-info.txt matches info in DB
needed to fix a couple other tests based on the stricter requirements
here. also changes the name.

We may want to consider moving some of these into a Chipmunk:: module.
We should probably move the specs for validation from bag_move_job to a
separate spec - the remnant there is from the history of these things
being combined in BagMoveJob.