Split Uploader#upload into two parts
This is backwards-compatible in case anything anywhere depends on the
`upload` method specifically. However, it is now also possible to upload
without waiting for the result, and you can wait for the result later.

If you wait for the result without first uploading, waiting for the
result won't do anything (it'll return nil). Otherwise, it returns true
or false in the same way as the original upload method does.